newt - don't automatically recompile if dependency is gone.
Old behavior: If a source file depends on a header (according to the .d
file), and that header has been deleted, newt recompiles the source
file.  This is problematic when the following sequence occurs:

1. Creates a new header file ("new.h")
2. An old header ("old.h") is modified to include the new header.
3. The project is built.  Now all source files which include "old.h"
   also depend on "new.h".
4. Delete "new.h".

Now on every build, all source files that include "old.h" get recompiled
unnecessarily.

Fix: When newt determines that a dependency has been deleted, it
rebuilds the .d file for the source file and restarts the process of
determining if the source file needs to be compiled.