Bug fix for -o option (log file).
Configure logging twice.  First just configure the filter for stderr;
second configure the logfile if there is one.  This needs to happen in
two steps so that the log level is configured prior to the attempt to
open the log file.  The correct log level needs to be applied to file
error messages.