newt - Don't link when there are no dep changes.
This fixes a bug that was introduced in
a996d13f7d971ad6f4d35a8d2ee066002c55b5ae

As a consequence of that change, the link step was performed for every
build, even if there were no changes among dependencies.  The problem
was that build flags from the compiler package were not applied to the
link command when it was checked against the previous link command.
This caused it to appear that the command line invocation had changed
since the previous link.