Mfg - collect instead of build; include img files
The "newt mfg create" command no longer builds the constituent targets.
Instead it expects the targets / images to have already been built.

Rationale: There is some extra configuration required for building
images: version numbers, build ID, and security material.  If the mfg
commands are responsible for creating images, an mfg configuration would
need to contain this information just so that it could forward it to the
create image command.