newt - Create manifests at build-time
Previously, newt only created a manifest when creating an image.  This
had the potential to be misleading, as the list of git commits didn't
necessarily reflect what went into the build.

Now, manifest creation is done in two steps:

1. Create manifest at build time ("newt build").

This initial manifest contains the following fields:
    * build_time
    * pkgs
    * loader_pkgs
    * target
    * repos

2. Augment manifest at image creation time.

The following fields are added to the existing manifest:
    * build_version
    * id
    * image
    * image_hash
    * loader
    * loader_hash

An "added bonus" is that boot loaders now have manifests.