newt - Allow "install" from local machine
This commit is to aid in testing of unreleased version of Mynewt.

Prior to his change, newt needed to clone a repo in order to determine
which branch or tag it should check out.

Now, a project.yml file can specify a repo type of "local".  When
installing from a local repo, newt assumes an up-to-date clone exists at
the specified path, so it just copies the directory rather than cloning
with git.

Example local repo (project.yml):

    repository.apache-mynewt-core:
        type: local
        vers: 0-latest
        path: "/Users/ccollins/repos/tmp/core"