MYNEWT-500 newt - preserve src dir tree in bin dir
This addresses an ambiguity that occurs when a package contains two
source files with the same filename.  For example, the arduino repo
(until very recently) contained the following two files:

    hw/mcu/atmel/samd21xx/src/sam0/utils/cmsis/samd21/source/gcc/startup_samd21.c
    hw/mcu/atmel/samd21xx/src/startup_samd21.c

Both of these files are in the same package (hw/mcu/atmel/samd21) and
they both have the same filename (startup_samd21c).

During a build, newt compiles both, but only links in one object file.

The problem is that the bin directory structure does not mirror that of
the source tree. Specifically, all of a package's object files get put
in the same directory; subtrees are not preserved. Consequently, the
second .o file overwrites the first.

The fix is to preserve the source directory structure in the bin
directory.