newt - Add missing O_TRUNC flags.
When newt overwrites an existing file, it needs to specify the
os.O_TRUNC flag to os.Open().  Otherwise, the file gets overwritten in
place and stale data might remain at the end.  This was causing
confusing issues with invalid dependency files (.d).

This issue is new because we no longer invoke the shell for external
commands.  Previously, we truncated files by redirecting shell output.