MYNEWT-503 newt - No status or log msgs at init
No log messages or status messages are displayed while the project is
being initialized.  This issue is related to tab completion: output is
intentionally suppressed during initialization so that the "complete"
command generates clean output.

The fix is to only generate valid tab completion values when it has been
determined that the "complete" command is being used.