newt - Continue reading pkgs on repo read error.
The process of populating the package list was aborting when there was a
repo read error.  This left newt with an incomplete package list.
Worse, the order that repos get read in is indeterminate, so the results
were inconsistent.

Now, newt prints a warning message when there is a repo read error, but
continues reading the remaining repositories.