MYNEWT-498 Newt - deps and syscfg calculated twice
This commit removes all duplicate syscfg calculation, and removes one
iteration of dependency calculation.

We still calculate dependencies a few times:
    (1) While calculating syscfg
    (2) While calculating loader deps (only for split)
    (3) While calculating app deps

We could combine (1) and (3) as an optimization.  (2) and (3) need to
remain separate, as the loader and app need to retain independent
dependendency graphs in order to identify common symbols during linking.