newt - Restructure dep resolution code
The dependency, API, and syscfg resolution code was bordering on
spaghetti; this commit restructures it to be simpler.  Initially, this
code was mostly sane, but it started to get out of control when the need
to calculate syscfg for split images arose.

Here is a sketch of the dependency resolution control flow:

1. First, calculate syscfg and determine which package provides each
required API.  Syscfg and APIs are project-wide; that is, they are
calculated across the aggregate of all app packages and loader packages
(if any).  The dependency graph for the entire set of packages gets
calculated here as a byproduct.

2. If there is no loader, then the set of all packages is just the app
packages.  We already resolved the necessary dependency information when
syscfg was calculated above, so we are done.

3.  Otherwise, we need to resolve dependencies separately for:
    a. The set of loader pacakges, and
    b.. The set of app pacakges.
These need to be resolved separately so that it is possible later to
determine which packages need to be shared between loader and app.