MYNEWT-506 newt - Compile files from base dir
Note: this change requires a sibling change to be made to core and the
other Mynewt repos.  Any package which specifies a path in its cflags
(e.g., "-I") will be broken by this change.  Rather than specify a path
relative to the source directory, the path should be relative to the
project base directory.