newt - Update min-commits due to compat-brkng chgs
Recently, newt changed the way it compiles source files.  Specifically,
it used to cd to each source directory when compiling; now it builds
everything from the project base directory.  Mynewt repos had to be
updated to accommodate this change: their -I cflags needed to specify
paths relative to the project directory.

This commit ensures the user isn't using a new newt tool with an old set
of repos.