newt: Prepare support for header skipping
In preparation for a tool to sign Zephyr images, add support for
Zephyr-style headers.

When Zephyr is configured to allow space for the image header
(CONFIG_TEXT_OFFSET), it's build system emits a binary with zeros padded
out for this header space.  To support this, add a `SrcSkip` option to
the `Image` struct to indicate a number of bytes of zeros to skip at the
beginning of a binary image.  Print a warning if this header doesn't
actually contain zeros, since this usually indicates that the parameter
is incorrect.

This patch adds the field to the struct, but since no code yet sets this
field, there should be no change in functionality.