newt: Prepare support for header padding
Since Zephyr is able to use the vector table directly out of flash, and
some MCUs have requirements on the alignment of this header, add a
`HeaderSize` value to the `Image` struct.

If this is set to a value larger than the `IMAGE_HEADER_SIZE`, padding
will be inserted between the header and the image.  The `HdrSz` field in
the header will be set accordingly, so that the image can still be
found.

This patch adds the field to the struct, but since no code yet sets this
field, there should be no change in functionality.