MYNEWT-558 newt - depgraph should show api / hard
The `target dep` and `target revdep` commands only indicated package
names in the dependency graph.  This change enhances the output of these
commands to show what created the dependency in the first place.
Specifically, they now show whether the dependency is "hard" (package
explicitly depends on another) or "soft" (package requires an API; other
package happens to implement that API).  This is useful when trying to
completely remove a package from a build.

In the future, we should also indicate which, if any, syscfg settings
generated a particular dependency.

The fix required a lot of changes to the newt code.  The major change
was to scale back the responsibilities of the pkg.LocalPackage type.
This struct was being used during build and depgraph resolution, and its
client relied on its inheritance of the pkg.Package interface.
Consequently, it was difficult to add new features without breaking
everything or bloating the LocalPackage struct.

Now, instances of LocalPackage get converted to resolve.ResolvePackage
during dependency and API resolution.  Furthermore, builder.BuildPackage
and resolve.ResolvePackage no longer inherit from LocalPackage.
Instead, each contains a pointer to its corresponding LocalPackage or
ResolvePackage.  Since these types never need to be treated as generic
package types, the code is simplified by eliminating these edges from
the inheritance tree.