MYNEWT-543 Keep settings not set in syscfg variable
The current newt target set target-name syscfg variable saves the settings in a new target syscfg.yml file. Settings from the current target syscfg.yml that are not specified in the syscfg variable are deleted. While setting a variable with only the values specified in the command is consistent with how other variables are set, the syscfg variable is a bit different because of multiple configuration settings.  A user might want to only override some settings but keep the values for the other settings in the target syscfg.yml.

This commit adds support to keep the values of the settings that are not specified in the set syscfg variable and to add or override the values in the target syscfg.yml for the settings specified for the variable. The code change converts the setting/value pair from the current syscfg.yml in to a KV String, and appends the KV String from command to it before it processes and saves the setting values. This way, an old setting value is either saved, or overridden if it is specified in the syscfg variable.

A user who only wants to create a new target syscfg.yml with the settings specified in the syscfg variable can set  syscfg="" before it sets syscfg with the setting names and values. Example:
newt target set target_name syscfg="" syscfg=settingname1=settingvalue1:settingname2=settingvalue2 ...