MYNEWT-753 newt - sysinit file inconsistent
In the generated sysinit C file, calls to initialization functions are
sorted by stage number. Within a stage number, the ordering is random,
and varies from one build to the next. The randomness comes from
Golang's random iteration of maps.

This is bad because it prevents repeatable builds, and causes
unnecessary rebuilds.

The fix is to sort alphabetically by package name within a stage number.

So,
    * First sort by stage number
    * Then sort by package name.