MYNEWT-640 added newt target amend command. newt target amend command allows user to add, modify, delete values for mulit-value variables such as syscfg, cflags, etc and addresses the limitation of the newt target set command raised in MYNEWT-543.
The amend command uses the same format as newt target set.  The variables that
can have values appended are: syscfg, cflags, aflags, and lfags.
It also supports a -d flag which allows the user to delete values from the
variable.

This replaces the previous commit for newt target append command:
9133a099e7906128d8825b4b4481dcb45e5a5480

Decided to call the command "amend" instead of "append" since it allows the user to modify current values, not just add to it.