image: Add support for RSA-PSS signatures
Currently, RSA signatures are done with PKCS#1 v1.5.  Newer versions of
this standard (starting with 2.1) define a newer signing algorithm, PSS,
and enourage adoption of this algorithm over the older method.

Add support for this to the 'newt' tool so that create-image can sign
RSA with PSS.  Since the keys are the same, add a command line argument
`--rsa-pss` to use the new algorithm.  Once this is adopted more fully
by the bootloader this could be made the default.