syscfg - true vals only in conditional overrides.
This commit fixes a bug involving conditional overrides of syscfg
values.  In the following syscfg listing:

    syscfg.vals.MY_SETTING:
        LOG_LEVEL: 255

LOG_LEVEL should only get set to 255 if MY_SETTING is defined with a
true value (nonzero and non-empty-string).  However, the actual behavior
is that LOG_LEVEL gets overridden if MY_SETTING is defined at all,
regardless of its value.

The fix is to only apply the override if the condition is true.