syscfg - Allow a pkg to override its own setting.
PKG-A can only override a setting defined by PKG-B if PKG-A has a
greater priority than PKG-B.  If PKG-A's priority is less than or equal
to PKG-B's, the override attempt is rejected and reported as a priority
violation.

This commit relaxes that rule slightly: a package can override settings
that it itself defines.

This behavior is needed  for making a setting's default value
conditional on another setting.  For example:

    syscfg.defs:
        TIMER_0:
            description: 'NRF52 Timer 0'
            value:  1
        TIMER_3:
            description: 'NRF52 Timer 3'
            value:  0

    # Use timer 3 instead of timer 0 if target wants to use the low
    # power clock.
    syscfg.vals.BLE_LP_CLOCK:
        TIMER_0: 0
        TIMER_3: 1