Add support for configuring the image header size
Some architectures require the image to be flashed at a specific
location, for example, k64f needs vectors to be at a multiple of 0x200
which makes it confusing to define partitions because it requires to
subtract the header size.

This patch makes it possible to configure the header size in target.yml
using the key "target.header_size" which will then, surprise!, write a header
with the requested size allowing the binary firmware to be located at a
well-known address.

Signed-off-by: Fabio Utzig <utzig@apache.org>