newt: Workaound incorrectly resolved symbols in 'newt size'
For unknown reason arm-none-eabi-nm does not resolve source files
properly for some symbols - it just leaves them without filename.
As a result 'newt size' put them in '(other)' group which is not
really helpful.

This patch adds workaround for this by calling 'addr2line' on symbols
with missing filename and this tool seems to have no problem with
finding proper filename.