newt - Don't trace stack during syscfg access
The purpose of this commit is to speed up newt.  During dependency
resolution of a target with a lot of syscfg settings (e.g., anything
using nimble), newt spent the majority of its time creating stack
traces.  These stack traces were created each time newt determined if a
syscfg setting was enabled.

The culprit was the `util.AtoiNoOct()` function.  This function attempts
to convert a string to an integer, allowing base-10 and base-16
representations.  If the string cannot be converted, a new `NewtError`
is returned.  Each time a `NewtError` is created, the current stack
trace is requested and placed in the error object.

The `AtoiNoOct()` function was being called repeatedly, as a syscfg
value must be compared to "0" to determine if it is enabled or not.

The fix is to introduce a new function: `AtoiNoOctTry()`.  This function
is similar to `AtoiNoOct()`, except it returns a bool indicating success
/ failure rather than a `NewtError`.

Example timings:

*Before fix*
```
[ccollins@ccollins-mac:~/repos/mynewt/core]$ time newt -lfatal -s target config
show bleprph-nrf52dk

real    0m1.052s
user    0m2.022s
sys     0m0.197s
```

*After fix*
```
[ccollins@ccollins-mac:~/repos/mynewt/core]$ time newt -lfatal -s target config
show bleprph-nrf52dk

real    0m0.301s
user    0m0.356s
sys     0m0.037s
```