newt: Add support for PKCS#8 private keys
Add support for RSA and ECDSA private keys that are represented in the
unencrypted PKCS#8 format.  This format is recommended for newer
applications, and can be generated by either newer versions of the
`imgtool.py` script from MCUboot, or by using the `genpkey` subcommand
of OpenSSL (instead of `rsa` or `ec`).  Other than the format better
capturing the type of key contained within it, this better prepares us
to be able to handle PKCS#8 keys that have been password protected.

This also adds some unit tests to verify that the code is able to load
the 4 supported key types:

   - RSA in "old" RSA private key format: (BEGIN RSA PRIVATE KEY)
   - ECDSA in "old" ECDSA private key format (BEGIN EC PRIVATE KEY)
   - RSA and ECDSA keys in PKCS#8 (BEGIN PRIVATE KEY)