Add support for Hash config in PBKDF2
Since PBKDF2 allows for any of a variety of hashes to be used and
imgtool delegates the algos used to pyca/cryptography the actual hash
used in the PBKDF2 can vary based on installation. The code was using
sha1 every time and this updates it to also support sha-224 and sha-256.

Signed-off-by: Fabio Utzig <utzig@apache.org>