Move repo version type to newtutil
This commit:

1. Removes the `interfaces.VersionInterface` and
   `interfaces.VersionReqInterface` types.

2. Moves `repo.Version`      to `newtutil.RepoVersion` and
         `repo.VersionMatch` to `newtutil.RepoVersionReq`

The version interfaces in the `interfaces` packages were only necessary
to work around Go's strict rules on package dependency cycles.  There
was only one implementation of each of these interfaces (`repo.Version`
and `repo.VersionMatch`), so the interfaces weren't solving a real
problem.  Since the `newtutil` package doesn't depend on any other newt
pacakges, it can contain the repo version types without risking circular
dependencies.

Doing away with the verison interfaces helps in a number of ways:

* Eliminates the need for reference semantics.  This allows for easy
comparison of map keys and a built-in copy operation.

* Eliminates the need to construct new containers when switching from
concrete to abstract types.  E.g., `[]*Version` and
[]interfaces.VersionInterface` are not compatible types, so a new slice
must be constructed and filled in to convert between the two.  This
problem goes away when there is only one version type.