Always copy local repos when fetching.
The "local" repo type is a debug feature.  It allows a project to
specify a local filepath for the repo remote rather than a git URL or
github specifier.  This feature is useful for testing changes to a
repo's `repository.yml` file before they are pushed.

There was a bug involving local repos: the repo directory was only
copied to the project on the initial install.  Subsequent upgrades would
not pick up changes to the source directory.