Repo dependencies
Allow a repo to specify which other repos it requires in its
`repository.yml` file.  For example, a `repository.yml` file might
contain the following:

```
    repo.deps:
        mynewt-nimble:
            type: git
            url: 'git@github.com:apache/mynewt-nimble.git'
            vers:
                master:             '0-dev'
                mynewt_1_4_0_tag:   '>=1.4.0'

        mynewt-nffs:
            type: git
            url: 'git@github.com:apache/mynewt-nffs.git'
            vers:
                master:             '0-dev'
                mynewt_1_4_0_tag:   '1.4.0'
```

What this says is:
    * The master branch of of this repo depends on:
        * nimble 0-dev
        * nffs 0-dev
    * The mynewt_1_4_0_tag tag of of this repo depends on:
        * nimble 1.4.0 or later
        * nffs 1.4.0 only

When the user performs an install, upgrade, or sync operation, newt
includes the depended-on repos in the operation.  For example, if the
user tells newt to install version 1.4.0 of the above repo, newt will
also install suitable versions of nimble and nffs.  If the user then
upgrades the above repo to 0.0.0, both nimble and nffs will be upgraded
to 0.0.0 along with it.

If the user's request cannot be completed due to conflicting repo
version requirements, newt informs the user of the problem and leaves
the project unchanged.

In addition, the `-a` (`--ask`) flag has been added to the install,
upgrade, and sync commands.  When this flag is specified, newt describes
the operation it is about to perform and prompts the user for
confirmation before proceeding.