Initialize and update submodules after checkout.
Always initialize and update submodules on checkout.  This is necessary
for repos that contain submodules.  Without this change, such repos can
be left in a modified "(new commits)" state immediately after switching
branches.  If the submodules have already been updated, this does not
generate any network activity.

For example, before this commit, when I try to install version 1.1.0 of
mcuboot with `newt upgrade`,

    [ccollins@ccollins-mac:~/tmp/myproj3/repos/mcuboot]$ git status
    On branch v1.1.0
    Changes not staged for commit:
      (use "git add <file>..." to update what will be committed)
      (use "git checkout -- <file>..." to discard changes in working directory)

            modified:   sim/mcuboot-sys/mbedtls (new commits)

    no changes added to commit (use "git add" and/or "git commit -a")

Subsequent update operations will fail due to the local modifications.