Allow package-level syscfg restrictions
Prior to this commit, syscfg restrictions could only be specified at the
setting level.  For example:

    syscfg.defs:
        LOG_CLI:
            value: 0
            restrictions:
                - SHELL_TASK

Sometimes, it is desirable to place restrictions on an entire package.
For example, the `hw/bsp/arduino_zero` package ought to require exactly
one of `BSP_ARDUINO_ZERO_PRO` or `BSP_ARDUINO_ZERO` be set.  Currently,
this restriction can only be imposed by creating a new setting for the
sole purpose of containing this restriction (and hoping no one overrides
the setting!), or by "shoehorning" the restriction into an existing
setting.  Another example where it would be useful to impose
restrictions at the level of the package is for apps.  An app may
require some settings to be enabled to function properly.

This commit allows restrictions to be defined at the top-level of a
`syscfg.yml` file.  For example:

    syscfg.defs:
        # ...

    syscfg.vals:
        # ...

    syscfg.restrictions:
        - 'BSP_ARDUINO_ZERO ^^ BSP_ARDUINO_ZERO_PRO'
        - SHELL_TASK