Don't generate syscfg for each package
This is an optimization: don't regenerate the entire syscfg for each
package.

For some packages, newt injects settings that should not be visible to
any other packages (e.g., SPLIT_APPLICATION).  When dealing with such
packages, newt makes a copy of the global syscfg and adds the injected
settings.

If a package has no injected settings, there is no need to copy the
global syscfg.  However, newt was making a copy for each package,
regardless of whether or not there were any injected settings.

This commit changes newt such that it only copies the global syscfg for
packages with injected settings.