syscfg - Allow a package to be deleted
This commit allows a package to be removed from the syscfg state after
its settings have been read.  This action is necessary in the case of
conditional dependency lists.  In this case, newt may require several
iterations of dependency resolution before discovering that a package
loaded early in the process has no dependents.

Deletion of package `p` from syscfg consists of the following steps:
    1. For each setting `s` defined by `p`:
        a. Delete `s` from the syscfg state.
        b. If `s` has any overrides, add them to the orphan list.
    2. For each setting `S` not defined by `p`:
        a. If `p` overrides `S`, remove the override from `s`'s override
           history