syscfg - Delay error detection until end
Prior to this commit, syscfg errors were detected as the syscfg state
was calculated.  This is problematic because some errors may naturally
be resolved as syscfg calculation progresses.  For example, a priority
violation (e.g., library overrides app's setting) may be detected, but
later it is discovered that the library package is actually not part of
the build.  By delaying error detection until syscfg is fully
calculated, this spurious error is not reported.