Allow git commit strings in `project.yml`
This fixes #157.

The `project.yml` file contains a set of repository descriptors, each
starting with `repository.`.  Each descriptor contains a `vers` field,
indicating the version of the repo that the project depends on.  Prior
to this commit, the `vers` field had to contain one of the following
forms:

    * [Normalized version]: #.#.#
        (e.g., "1.3.0")

    * [Floating version]:   #[.#]-<stability
        (e.g., "0-dev")

This commit adds support for a third form:

    * [Git commit]:         <git-commit-ish>-commit
        (e.g., "0aae710654b48d9a84d54de771cc18427709df7d-commit")

In addition, this commit removes the ability to "force update" and
"force sync" (i.e., the `-f` switch no longer has any effect with these
commands).