Ensure new .a file needed before deleting old
When creating a .a file for a package, newt first deletes the existing
.a file from the previous build.  There is one case where this causes
problems: the package's source directory contains a .a file with the
same name that newt plans to use.  In this case, newt confuses the
input .a file as an artifact from a prior build and deletes it.

The fix is to delay the deletion of the old .a file until after ensuring
that a new .a file will be built.  Newt determines if it will be
building a new .a file by checking if any .o files were produced for the
package.  Thus, a package consisting only of .a files does not generate
a new .a file.