Fix `install -f` for non-master repo commits
If the `project.yml` file specified a repo version that maps to a tag,
then force installs (`install -f`) consistently failed with a message
similar to the following:

    Error: Error updating "apache-mynewt-nimble": Cannot determine commit type of "nimble_1_0_0_tag"

The problem was caused by this sequence of events:

1. Newt fetches latest commits, branches, and tags.
2. Newt deletes old repo (due to `-f` flag).
3. Newt clones repo again.
4. Newt attempts to checkout the desired commit.

Step 4 failed because newt never re-fetched the list of tags.  The fix
is to retrieve the list of tags between steps 3 and 4.