Fix precedence of binary operators
Newt was using an unintuitive set of operator precedences for binary
operators.  In particular, `==` and `!=` had higher precedence than the
other operators.  This led to surprising behavior like:

    LOG_LEVEL == 1 || LOG_VERSION == 3

being parsed as:

    LOG_LEVEL == (1 || (LOG_VERSION == 3))

when what the user obviously wanted is:

    (LOG_LEVEL == 1) || (LOG_VERSION == 3)

This breaks backwards compatibility in some sense, but I can't imagine
that anyone is relying on the current broken behavior.  Anyone who wants
the expected behavior is already enforcing it with extra parentheses, so
this change won't affect them.