Eliminate unnecessary linker invocations
Somtimes newt would re-link the .elf file when no link was necessary.
This happened because:

(1) Newt specified all packages' cflags and lflags in the linker command.
(2) When collecting package flags, newt would iterate the list
of packages in an unpredictable order.

When invoking the linker, newt sorts the flags alphabetically, so (2) by
itself is not an issue.  However, if two or more flags are in conflict,
newt resolves the conflict by discarding all but the first.  For example,
newt collapses the following:

    -std=c99 -std=c89 -std=gnu99

into

    -std=c99

Since the flags are presented in an unpredictable order, the final form
of the command is inconsistent from one run to the next.  If newt
detects a change in command line since the last linker invocation, it
decides that an additional link is required.  Worse, different flags are
passed to the linker each time! (though, luckily, it seems these flags
are ignored by the linker since they are compiler flags, at least those
specified by the apache-mynewt-core packages).

The fixes are:

1. When building the linker command, only extract each package's
*lflags* (not cflags).
2. Iterate the package list in a predictable order (alphabetically).