Clean up downloader API - Remove "commit" state
This commit removes two functions from the Downloader API:

    * GetCommit
    * SetCommit

These functions didn't actually query the git repo.  Instead, they
manipulated a cached commit string that would be use in subsequent
operations.

The cached commit was only used in one place.  It is simpler to just
specify the commit as an argument as needed.