Upgrade: Ensure "origin" points to correct URL
On `newt install`, `newt upgrade`, and `newt sync`, force "origin"
remote to point to the URL specified in `project.yml` (or in another
repo's dependency specification, if the repo is not described in
`project.yml`).

This is necessary when the user changes `project.yml` to have a
different URL for an already-installed repo.  This would happen if the
user is switching between github forks.