Don't start an upgrade or sync if repos are dirty
Before starting an upgrade or sync operation, check if any repos being
operated on are in a dirty state.  If any repo is dirty, abort with a
descriptive message.  Optionally, the user can specify `-f` (force) to
try the operation anyway.

A repo is dirty if any of the following is true:
    * Repo contains local changes
    * Repo contains staged changes
    * Repo contains unpushed commits