Improve `newt info` command; show repo ver
This commit adds some information to the output of the `newt info`
command.  If no arguments are specified, the command lists all repos in
the project, and appends the following flags as appropriate:

    * Upgrade needed
    * Dirty: <description of dirty state>

If the user provides an argument, the command's old behavior is
retained.  I.e., the command prints a list of all packages belonging to
the specified repo.