Improve `newt info` command; show repo ver
This commit adds some information to the output of the `newt info`
command.  If no arguments are specified, the command lists all repos in
the project, and appends the following flags as appropriate:

    * Currently installed version (or "not installed")
    * Dirty: <description of dirty state>

In addition, if the user specifies the `-r` (remote) switch, the command
fetches the latest `repository.yml` files and determines if any repos
can be upgraded.

If the user provides an argument, the command's old behavior is
retained.  I.e., the command prints a list of all packages belonging to
the specified repo.  This behavior is retained for backwards
compatibility.